new_tests/tests_requirements.txt
ubirch-protocol
pytest

msgpack>=0.6.0
ecdsa>=0.7
ed25519>=1.4